Fix for issue #22
Fix bug where parties with no 'now playing' updates,
such as bbc6music, weren't scrobbling. This was
because we were relying on the time of 'now playing'
events for the scrobble timestamp. When this happens
we now use the current time and hope that this
gets past the last.fm filters.